Stop overriding file structure on "docker-compose up".

+++ 10.0-php5.6-apache/docker-entrypoint.sh
@@ -2,15 +2,19 @@
 
 set -euo pipefail
 
-# extracting
-unzip dle.zip upload/*
-mv upload/* /var/www/html
+# directory is empty
+if [ -z "$(ls -A /var/www/html)" ]; then
 
-# clearing
-rm dle.zip
-rm -rf upload
+    # extracting
+    unzip dle.zip upload/*
+    mv upload/* /var/www/html
 
-# define owner
-chown -R www-data:www-data /var/www/html
+    # clearing
+    rm dle.zip
+    rm -rf upload
+
+    # define owner
+    chown -R www-data:www-data /var/www/html
+fi
 
 exec "$@"
